+++ CommHandler/src/main/java/com/smartgridready/communicator/common/api/values/Int64UValue.java
@@ -1,7 +1,11 @@
 package com.smartgridready.communicator.common.api.values;
 
 import java.math.BigInteger;
+import java.time.Instant;
 import java.util.Map;
+
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.node.BigIntegerNode;
 
 public class Int64UValue extends Value {
 
@@ -88,6 +92,16 @@ public class Int64UValue extends Value {
     @Override
     public Map<String, Boolean> getBitmap() {
         throw new UnsupportedOperationException("Cannot convert from int64U value to a bitmap value.");
+    }
+
+    @Override
+    public Instant getDateTime() {
+        return Instant.ofEpochMilli(getInt64());
+    }
+
+    @Override
+    public JsonNode getJson() {
+        return BigIntegerNode.valueOf(value);
     }
 
     @Override