+++ CommHandler/src/main/java/com/smartgridready/communicator/common/api/values/Value.java
@@ -4,9 +4,15 @@ import com.smartgridready.ns.v0.DataTypeProduct;
 import com.smartgridready.ns.v0.EnumMapProduct;
 import com.smartgridready.ns.v0.ModbusBoolean;
 import com.smartgridready.ns.v0.ModbusDataType;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.smartgridready.communicator.common.helper.JsonHelper;
 import com.smartgridready.communicator.modbus.helper.ConversionHelper;
+import com.smartgridready.driver.api.common.GenDriverException;
 
 import java.math.BigInteger;
+import java.time.Instant;
 import java.util.Map;
 
 public abstract class Value  {
@@ -25,10 +31,23 @@ public abstract class Value  {
     public abstract boolean getBoolean();
     public abstract EnumRecord getEnum();
     public abstract Map<String, Boolean> getBitmap();
+    public abstract Instant getDateTime();
+    public abstract JsonNode getJson();
     public abstract void absValue();
     public abstract void roundToInt();
 
     public abstract Value[] asArray();
+
+    public <T> T getJson(Class<T> aClass) {
+        JsonNode node = getJson();
+        ObjectMapper mapper = new ObjectMapper();
+        try {
+            return mapper.treeToValue(node, aClass);
+        } catch (JsonProcessingException e) {
+            var msg = String.format("Unable to map JSON node to the given class '%s'", aClass.getSimpleName());
+            throw new IllegalArgumentException(msg);
+        }
+    }
 
     public int[] toModbusRegister(ModbusDataType modbusDataType) {
 
@@ -68,6 +87,9 @@ public abstract class Value  {
         if (modbusDataType.getBoolean() != null) {
             short booleanAsShort = mapBooleanToShort(modbusDataType.getBoolean(), getBoolean());
             return ConversionHelper.shortToRegister(booleanAsShort);
+        }
+        if (modbusDataType.getDateTime() != null) {
+            return ConversionHelper.longToRegisters(getInt64());
         }
 
         throw new IllegalArgumentException(
@@ -125,6 +147,9 @@ public abstract class Value  {
             short booleanAsShort = ConversionHelper.byteBufFromRegisters(registers).getShort();
             return BooleanValue.of(mapShortToBoolean(modbusDataType.getBoolean(), booleanAsShort));
         }
+        if (modbusDataType.getDateTime() != null) {
+            return DateTimeValue.of(Instant.ofEpochMilli(ConversionHelper.byteBufFromRegisters(registers).getLong()));
+        }
 
         throw new IllegalArgumentException(String.format("Modbus register type %s to Value.class conversion from register not supported.",
                 DataType.getModbusDataTypeName(modbusDataType)));
@@ -168,6 +193,16 @@ public abstract class Value  {
         }
         if (dataType.getBoolean() != null) {
             return BooleanValue.of(Boolean.parseBoolean(value));
+        }
+        if (dataType.getDateTime() != null) {
+            return DateTimeValue.of(Instant.parse(value));
+        }
+        if (dataType.getJson() != null) {
+            try {
+                return JsonHelper.parseJsonResponse(null, value);
+            } catch (GenDriverException e) {
+                throw new IllegalArgumentException("Could not convert string to JSON", e);
+            }
         }
 
         throw new IllegalArgumentException(String.format("Generic type %s conversion from String to Value.class conversion from register not supported.",