+++ CommHandler/src/main/java/com/smartgridready/communicator/messaging/impl/SGrMessagingDevice.java
@@ -15,9 +15,11 @@ import com.smartgridready.ns.v0.MessagingDataPoint;
 import com.smartgridready.ns.v0.MessagingDataPointConfiguration;
 import com.smartgridready.ns.v0.MessagingFunctionalProfile;
 import com.smartgridready.ns.v0.MessagingInterface;
+import com.smartgridready.ns.v0.MessagingValueMapping;
 import com.smartgridready.ns.v0.OutMessage;
 import com.smartgridready.ns.v0.ResponseQuery;
 import com.smartgridready.ns.v0.ResponseQueryType;
+import com.smartgridready.ns.v0.ValueMapping;
 import com.smartgridready.communicator.common.api.values.StringValue;
 import com.smartgridready.communicator.common.api.values.Value;
 import com.smartgridready.communicator.common.helper.JsonHelper;
@@ -29,7 +31,9 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.util.Collections;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -108,17 +112,21 @@ public class SGrMessagingDevice extends SGrDeviceBase<
 
     @Override
     public Value getVal(String profileName, String dataPointName) throws GenDriverException {
-        return getVal(profileName, dataPointName, SYNC_READ_TIMEOUT_MSEC);
+        return getVal(profileName, dataPointName, null, SYNC_READ_TIMEOUT_MSEC);
     }
 
     @Override
     public Value getVal(String profileName, String dataPointName, Properties parameters) throws GenDriverException {
-        // parameters not supported, just return getVal()
-        return getVal(profileName, dataPointName);
+        return getVal(profileName, dataPointName, parameters, SYNC_READ_TIMEOUT_MSEC);
     }
 
     @Override
-    public Value getVal(String profileName, String dataPointName, long timeoutMs)
+    public Value getVal(String profileName, String dataPointName, long timeoutMs) throws GenDriverException {
+        return getVal(profileName, dataPointName, null, timeoutMs);
+    }
+
+    @Override
+    public Value getVal(String profileName, String dataPointName, Properties parameters, long timeoutMs)
             throws GenDriverException {
 
         MessagingDataPoint dataPoint = findDatapoint(profileName, dataPointName);
@@ -136,12 +144,23 @@ public class SGrMessagingDevice extends SGrDeviceBase<
                 .map(MessagingDataPointConfiguration::getInMessage)
                 .map(InMessage::getFilter).orElse(null);
 
-
-
         if (outReadCmdTopicOpt.isPresent()) {
+            Properties substitutions = new Properties();
+            // substitute default values of dynamic parameters
+            if (null != dataPoint.getDataPoint().getParameterList()) {
+				dataPoint.getDataPoint().getParameterList().getParameterListElement().forEach(p -> {
+					String pVal = (null != p.getDefaultValue()) ? p.getDefaultValue() : "";
+					substitutions.put(p.getName(), pVal);
+				});
+			}
+            // substitute actual request parameters
+			if (parameters != null) {
+				substitutions.putAll(parameters);
+			}
+
             // Read value from device
             OutMessage outMessage = outReadCmdTopicOpt.get();
-            return getValueFromDevice(timeoutMs, dataPoint, outMessage.getTopic(), outMessage.getTemplate(), inMessageTopic, inMessageFilter);
+            return getValueFromDevice(substitutions, timeoutMs, dataPoint, outMessage.getTopic(), outMessage.getTemplate(), inMessageTopic, inMessageFilter);
         } else {
             // Read value from cache
             MessageCacheRecord cacheRecord =  messageCache.get(MessageCacheKey.of(inMessageTopic, inMessageFilter));
@@ -152,7 +171,7 @@ public class SGrMessagingDevice extends SGrDeviceBase<
         }
     }
 
-    private Value getValueFromDevice(long timeoutMs, MessagingDataPoint dataPoint, String outMessageTopic, String outMessageTemplate, String inMessageTopic, MessageFilter messageFilter) throws GenDriverException {
+    private Value getValueFromDevice(Properties parameters, long timeoutMs, MessagingDataPoint dataPoint, String outMessageTopic, String outMessageTemplate, String inMessageTopic, MessageFilter messageFilter) throws GenDriverException {
         if (messagingClient == null) {
             throw new GenDriverException(NOT_CONNECTED);
         }
@@ -161,7 +180,7 @@ public class SGrMessagingDevice extends SGrDeviceBase<
 
         Either<Throwable, Message> result = messagingClient.readSync(
                 outMessageTopic,
-                Message.of(outMessageTemplate),
+                Message.of(replacePropertyPlaceholders(outMessageTemplate, parameters)),
                 inMessageTopic,
                 messageFilterHandler,
                 timeoutMs
@@ -177,17 +196,25 @@ public class SGrMessagingDevice extends SGrDeviceBase<
                     .map(MessagingDataPointConfiguration::getInMessage)
                     .map(InMessage::getResponseQuery);
 
+            Value value;
             if (queryOpt.isPresent()) {
                 ResponseQuery responseQuery = queryOpt.get();
                 if (responseQuery.getQueryType() != null && ResponseQueryType.JMES_PATH_EXPRESSION == responseQuery.getQueryType()) {
-                    return JsonHelper.parseJsonResponse(responseQuery.getQuery(), response);
+                    value = JsonHelper.parseJsonResponse(responseQuery.getQuery(), response);
                 } else if (responseQuery.getQueryType() != null && ResponseQueryType.JMES_PATH_MAPPING == responseQuery.getQueryType()) {
-                    return JsonHelper.mapJsonResponse(responseQuery.getJmesPathMappings(), response);
+                    value = JsonHelper.mapJsonResponse(responseQuery.getJmesPathMappings(), response);
                 } else if (responseQuery.getQueryType() != null) {
                     throw new GenDriverException("Response query type " + responseQuery.getQueryType().name() + " not supported yet");
+                } else {
+                    throw new GenDriverException("Response query type missing");
                 }
+            } else {
+                // mapping device -> generic (only for plain string values)
+                value = getMappedGenericValue(dataPoint.getMessagingDataPointConfiguration(), response);
             }
-            return StringValue.of(response);
+
+            // unit conversion before returning to client
+            return applyUnitConversion(dataPoint, value, SGrDeviceBase::multiply);
         } else {
             throw new GenDriverException(result.getLeft());
         }
@@ -213,8 +240,14 @@ public class SGrMessagingDevice extends SGrDeviceBase<
                 .map(OutMessage::getTemplate)
                 .orElseThrow(() -> new IllegalArgumentException("W and RW data-points need an outMessageTemplate to send the read command within EI-XML"));
 
-        // noinspection RegExpRedundantEscape
-        outMessageTemplate = outMessageTemplate.replace("{{value}}", value.getString());
+        // unit conversion before output
+        value = applyUnitConversion(dataPoint, value, SGrDeviceBase::divide);
+
+        // mapping generic -> device
+        String outValue = getMappedDeviceValue(dataPoint.getMessagingDataPointConfiguration(), value);
+
+        // no regex here, string literal replacement is sufficient
+        outMessageTemplate = outMessageTemplate.replace("[[value]]", outValue);
 
         messagingClient.sendSync(outMessageTopic, Message.of(outMessageTemplate));
     }
@@ -273,10 +306,14 @@ public class SGrMessagingDevice extends SGrDeviceBase<
                         throw new GenDriverException("Response query type " + queryOpt.get().getQueryType().name() + " not supported yet");
                 }
             } else {
-                value = StringValue.of(response);
+                // mapping device -> generic (only for plain string values)
+                value = getMappedGenericValue(dataPoint.getMessagingDataPointConfiguration(), response);
             }
             LOG.debug("Received subscribed message on topic={}, filter={}, payload={}",
                     inMessageTopic, queryOpt.isPresent() ? queryOpt.get().getQuery() : "none", response);
+
+            // unit conversion before inserting into cache
+            value = applyUnitConversion(dataPoint, value, SGrDeviceBase::multiply);
 
             messageCache.put(MessageCacheKey.of(inMessageTopic, messageFilter), MessageCacheRecord.of(value));
             callbackFunction.accept(Either.right(value));
@@ -361,4 +398,50 @@ public class SGrMessagingDevice extends SGrDeviceBase<
         });
         return count.get();
     }
+
+    private static Value getMappedGenericValue(MessagingDataPointConfiguration dataPointConfiguration, String value) {
+        String mappedValue = value;
+
+        List<ValueMapping> valueMappings = Optional.ofNullable(dataPointConfiguration)
+            .map(MessagingDataPointConfiguration::getInMessage)
+            .map(InMessage::getValueMapping)
+            .map(MessagingValueMapping::getMapping).orElse(Collections.emptyList());
+        for (ValueMapping mapping: valueMappings) {
+            if (mappedValue.equals(mapping.getDeviceValue())) {
+                mappedValue = mapping.getGenericValue();
+                break;
+            }
+        }
+
+        return StringValue.of(mappedValue);
+    }
+
+    private static String getMappedDeviceValue(MessagingDataPointConfiguration dataPointConfiguration, Value value) {
+        String mappedValue = value.getString();
+
+        List<ValueMapping> valueMappings = Optional.ofNullable(dataPointConfiguration)
+            .map(MessagingDataPointConfiguration::getWriteCmdMessage)
+            .map(OutMessage::getValueMapping)
+            .map(MessagingValueMapping::getMapping).orElse(Collections.emptyList());
+        for (ValueMapping mapping: valueMappings) {
+            if (mappedValue.equals(mapping.getGenericValue())) {
+                mappedValue = mapping.getDeviceValue();
+                break;
+            }
+        }
+
+        return mappedValue;
+    }
+
+    private static String replacePropertyPlaceholders(String template, Properties properties) {
+		// this is for dynamic parameters
+		String convertedTemplate = template;
+		if (template != null && properties != null) {
+			for (Map.Entry<Object, Object> entry : properties.entrySet()) {
+				// no regex here, string literal replacement is sufficient
+				convertedTemplate = convertedTemplate.replace("[[" + (String)entry.getKey() + "]]", (String)entry.getValue());
+			}
+		}
+		return convertedTemplate;
+	}
 }