+++ CommHandler/src/main/java/com/smartgridready/communicator/rest/http/client/RestServiceClient.java
@@ -42,6 +42,7 @@ public class RestServiceClient {
 	private static final ObjectMapper objectMapper = new ObjectMapper();
 
 	private final String baseUri;
+	private final boolean verifyCertificate;
 
 	private final RestApiServiceCall restServiceCall;
 	private final GenHttpClientFactory httpClientFactory;
@@ -55,12 +56,13 @@ public class RestServiceClient {
 		HTTP_METHOD_MAP.put(HttpMethod.DELETE, com.smartgridready.driver.api.http.HttpMethod.DELETE);
 	}
 
-	protected RestServiceClient(String baseUri, RestApiServiceCall serviceCall, GenHttpClientFactory httpClientFactory) throws IOException {
-		this(baseUri, serviceCall, httpClientFactory, new Properties());
+	protected RestServiceClient(String baseUri, boolean verifyCertificate, RestApiServiceCall serviceCall, GenHttpClientFactory httpClientFactory) throws IOException {
+		this(baseUri, verifyCertificate, serviceCall, httpClientFactory, new Properties());
 	}
 
-	protected RestServiceClient(String baseUri, RestApiServiceCall serviceCall, GenHttpClientFactory httpClientFactory, Properties substitutions) throws IOException {
-		this.baseUri = replacePropertyPlaceholders(baseUri, substitutions);
+	protected RestServiceClient(String baseUri, boolean verifyCertificate, RestApiServiceCall serviceCall, GenHttpClientFactory httpClientFactory, Properties substitutions) throws IOException {
+		this.baseUri = baseUri;
+		this.verifyCertificate = verifyCertificate;
 		this.restServiceCall = cloneRestServiceCallWithSubstitutions(serviceCall, substitutions);
 		this.httpClientFactory = httpClientFactory;
 	}
@@ -75,11 +77,11 @@ public class RestServiceClient {
 	}
 
 	public String getBaseUri() {
-		return getBaseUri(new Properties());
+		return baseUri;
 	}
 
-	public String getBaseUri(Properties substitutions) {
-		return replacePropertyPlaceholders(baseUri, substitutions);
+	public boolean isVerifyCertificate() {
+		return verifyCertificate;
 	}
 
 	public RestApiServiceCall getRestServiceCall() {
@@ -90,7 +92,7 @@ public class RestServiceClient {
 
 		var serviceCall = cloneRestApiServiceCall(restServiceCall);
 
-		// Substitutions can appear within the request path, request body or even the response query.
+		// Substitutions can appear within the request path, request headers, request body or even the response query.
 		serviceCall.setRequestPath(replacePropertyPlaceholders(serviceCall.getRequestPath(), substitutions));
 		serviceCall.setRequestBody(replacePropertyPlaceholders(serviceCall.getRequestBody(), substitutions));
 
@@ -125,7 +127,7 @@ public class RestServiceClient {
 		}
 
 		RestApiServiceCall serviceCall = getRestServiceCall();
-		GenHttpRequest httpRequest = httpClientFactory.createHttpRequest();
+		GenHttpRequest httpRequest = httpClientFactory.createHttpRequest(verifyCertificate);
 
 		httpRequest.setHttpMethod(mapHttpMethod(serviceCall.getRequestMethod()));
 		try {
@@ -181,23 +183,23 @@ public class RestServiceClient {
 	}
 
 	private static String replacePropertyPlaceholders(String template, Properties properties) {
-
+		// this is for dynamic parameters
 		String convertedTemplate = template;
 		if (template != null && properties != null) {
 			for (Map.Entry<Object, Object> entry : properties.entrySet()) {
-				// noinspection RegExpRedundantEscape
-				convertedTemplate = convertedTemplate.replaceAll("\\{\\{" + entry.getKey() + "\\}\\}", (String)entry.getValue());
+				// no regex here, string literal replacement is sufficient
+				convertedTemplate = convertedTemplate.replace("[[" + (String)entry.getKey() + "]]", (String)entry.getValue());
 			}
 		}
 		return convertedTemplate;
 	}
 
-	public static RestServiceClient of(String baseUri, RestApiServiceCall serviceCall, GenHttpClientFactory httpClientFactory) throws IOException {
-		return new RestServiceClient(baseUri, serviceCall, httpClientFactory);
+	public static RestServiceClient of(String baseUri, boolean verifyCertificate, RestApiServiceCall serviceCall, GenHttpClientFactory httpClientFactory) throws IOException {
+		return new RestServiceClient(baseUri, verifyCertificate, serviceCall, httpClientFactory);
 	}
 
-	public static RestServiceClient of(String baseUri, RestApiServiceCall serviceCall, GenHttpClientFactory httpClientFactory, Properties substitutions) throws IOException {
-		return new RestServiceClient(baseUri, serviceCall, httpClientFactory, substitutions);
+	public static RestServiceClient of(String baseUri, boolean verifyCertificate, RestApiServiceCall serviceCall, GenHttpClientFactory httpClientFactory, Properties substitutions) throws IOException {
+		return new RestServiceClient(baseUri, verifyCertificate, serviceCall, httpClientFactory, substitutions);
 	}
 
 	private static com.smartgridready.driver.api.http.HttpMethod mapHttpMethod(HttpMethod httpMethod) throws IOException {
@@ -206,7 +208,9 @@ public class RestServiceClient {
 	}
 
 	private static RestApiServiceCall cloneRestApiServiceCall(RestApiServiceCall restApiServiceCall) throws IOException {
-		var clone = objectMapper.writeValueAsString(restApiServiceCall);
-		return objectMapper.readValue(clone, RestApiServiceCall.class);
+		return objectMapper.readValue(
+			objectMapper.writeValueAsString(restApiServiceCall),
+			RestApiServiceCall.class
+		);
 	}
 }