+++ CommHandler/CHANGELOG.md
@@ -5,6 +5,50 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 2025-04-15
+
+### Fixed
+
+- fixed configuration parameter replacement in EID content when certain characters are used
+
+### Added
+
+- implemented value mappings in messaging interface similar to REST interface
+- added option to disable SSL certificate validation to REST interface
+- implemented introspection methods of dynamic data point parameters
+
+### Changed
+
+- requires updated SGr specification
+- data point "{{value}}" placeholders must be replaced with "\[\[value\]\]"
+
+
+## [2.3.0] - 2025-03-13
+
+### Fixed
+
+- fixed unit conversion in messaging
+
+### Added
+
+- added support for datetime data type (as ISO-8601 string or 64bit time stamp)
+- added support for json data type (with JsonNode and conversion to objects)
+
+### Changed
+
+- improved internal processing of JSON data
+
+
+## [2.2.0] - 2025-02-10
+
+### Changed
+
+- uses updated Modbus driver API
+- adapted Modbus implementation to new API
+- changed 'easymodbus' snapshot parameter to 'sgr-driver-modbus'
+- j2mod is new default Modbus driver
+
+
 ## [2.1.2] - 2025-01-06
 
 ### Changed