+++ GenDriverAPI/CHANGELOG.md
@@ -5,6 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] - 2025-04-14
+
+### Added
+
+- added option to disable SSL certificate validation to REST interface
+
+
+## [2.2.0] - 2025-02-10
+
+### Added
+
+- improved Modbus interface methods
+- support Modbus RTU with ASCII encoding
+
+### Changed
+
+- deprecated 'old' Modbus interface methods
+
+
 ## [2.1.0] - 2024-11-28
 
 ### Added